Hot garbage (#9)
* add instance to application modification

server instance in applications modified so that the server is better specified.

* Update default permissions in application

the application now includes the guild info and view craftbook grants

* move testing queue information

The testing queues do not serve a purpose and cause more confusion then necessary. Information about them have been moved to the FAQ section instead.

* grammatical change to result

NotInGuild tells you that it is an implied guild.